requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.1
fsspec==2024.9.0
gradio==4.42.0
gradio-client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.7
importlib-resources==6.4.4
jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.0
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
roboflow==1.1.44
ruff==0.6.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
